Break slides into multiple files
In order to use the slide content from Open edX, we need each slide (a
"vertical" in Open edX parlance) in its own separate Markdown file.

Split the original slides.md up into several files in the slides/
subdirectory, and create slides.md via a Makefile target.

This has the added benefit that we can reorder the slides as needed,
by simply rearranging the order in which the slides appear in the
Makefile. We can, if necessary, also order them differently in the
course and in the slide deck.