Break out HTML files
Due to some stupidity in HTML import in Open edX, we can't work with
inline processed HTML if we want code syntax highlighting to work. We
can fix this by using a silly workaround and tossing the HTML into a
separate file, which un-breaks syntax highlighting.

See hastexo/olx-utils#9 for a discussion of the underlying issue.